cube-builder-aws/requirements.txt
Flask>=1.0.3
Flask-JWT>=0.3.2
Flask-Script==2.0.6
Flask-Migrate>=2.5.2
Flask-SQLAlchemy>=2.4.1
GeoAlchemy2>=0.6.2
SQLAlchemy[postgresql]>=1.3.10
git+https://github.com/brazil-data-cube/bdc-db@master
git+https://github.com/brazil-data-cube/stac.py@master
psycopg2-binary==2.8.4
cerberus==1.3.1
boto3
botocore==1.13.43
numpy==1.18.0
rasterio==1.1.2
requests==2.22.0
numpngw==0.0.8